remove include js and rename request of update categories
